chore(folder-structure): moves other components into components folder
